+++ doc/release-notes.md
@@ -6,6 +6,7 @@
 - Added rate lookup helper when creating offer.
 - Prevent old shutdown link from shutting down a new session.
 - ui: Connected XMR wallet to rpc page.
+- Separate chain to generate smsg addresses.
 
 
 0.0.26